+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2026-07-17
+
+### Added
+
+- Koa framework support
+- API docs choices: Swagger UI, Scalar, OpenAPI JSON
+- Test runners: Vitest, Jest, Mocha + Chai
+
+### Changed
+
+- `features.apiDocs` / boolean `testing` replaced by `features.docs` and `features.testing` enums
+
 ## [1.0.0] - 2026-07-16
 
 ### Added